refactor: move the contents of `switch.rs`
